app/vmui: move handler to own file, fix show dashboards,
